+++ backend/tests/graphql/conftest.py
@@ -12,16 +12,45 @@ from pathlib import Path
 from sqlite3 import Connection
 
 import pytest
+import quart
+from ariadne import graphql
 
 from src.constants import Constants
 from src.enums import ArtistRole, CollectionType, PlaylistType, ReleaseType
+from src.graphql import error_formatter, schema
 from src.library import collection
 from src.library import playlist_entry as pentry
-from src.util import freeze_database_time
+from src.library import user
+from src.util import database, freeze_database_time
+from src.webserver.routes.graphql import GraphQLContext
 from tests.conftest import SEED_DATA
 from tests.factory import Factory
+from tests.fragments import FRAGMENTS
 
 GQL_DB_PATH = SEED_DATA / "gql_db.sqlite3"
+
+
+@pytest.fixture
+def graphql_query(seed_data, quart_app):
+    async def executor(query):
+        used_fragments = "\n".join(v for k, v in FRAGMENTS.items() if k in query)
+        query = f"{query}\n{used_fragments}"
+
+        async with quart_app.test_request_context("/testing", method="GET"):
+            with database() as conn:
+                return await graphql(
+                    schema=schema,
+                    data={"operationName": None, "variables": {}, "query": query},
+                    context_value=GraphQLContext(
+                        user=user.from_id(1, conn),  # type: ignore
+                        db=conn,
+                        request=quart.request,
+                    ),
+                    error_formatter=error_formatter,
+                    debug=False,
+                )
+
+    yield executor
 
 
 @pytest.fixture(scope="session")
